refactor : ProblemSearchCondition 위치 변경 ( domain 계층 - model )
