+++ src/main/java/org/ezcode/codetest/domain/problem/repository/ProblemRepository.java
@@ -2,7 +2,7 @@ package org.ezcode.codetest.domain.problem.repository;
 
 import java.util.Optional;
 
-import org.ezcode.codetest.application.problem.dto.request.ProblemSearchCondition;
+import org.ezcode.codetest.domain.problem.model.ProblemSearchCondition;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
