+++ src/main/java/org/ezcode/codetest/domain/problem/service/ProblemDomainService.java
@@ -1,12 +1,11 @@
 package org.ezcode.codetest.domain.problem.service;
 
-import org.ezcode.codetest.application.problem.dto.request.ProblemSearchCondition;
+import org.ezcode.codetest.domain.problem.model.ProblemSearchCondition;
 import org.ezcode.codetest.domain.problem.exception.ProblemException;
 import org.ezcode.codetest.domain.problem.exception.code.ProblemExceptionCode;
 import org.ezcode.codetest.domain.problem.model.ProblemInfo;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
 import org.ezcode.codetest.domain.problem.model.entity.ProblemSearchDocument;
-import org.ezcode.codetest.domain.problem.model.enums.Category;
 import org.ezcode.codetest.domain.problem.repository.ProblemDocumentRepository;
 import org.ezcode.codetest.domain.problem.repository.ProblemRepository;
 import org.springframework.data.domain.Page;
