+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/problem/ProblemRepositoryCustom.java
@@ -1,6 +1,6 @@
 package org.ezcode.codetest.infrastructure.persistence.repository.problem;
 
-import org.ezcode.codetest.application.problem.dto.request.ProblemSearchCondition;
+import org.ezcode.codetest.domain.problem.model.ProblemSearchCondition;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
